edit

+++ NetFlow.Blazor/NetFlow.Blazor.Shared/Models/ShippableOrdersFilterRequest.cs
@@ -9,7 +9,7 @@ namespace NetFlow.Blazor.Shared.Models
         public string Customer { get; set; } = string.Empty;
         public DateTime? StartDate { get; set; }
         public DateTime? EndDate { get; set; }
-        public int? Warehouse { get; set; }
+        public short? Warehouse { get; set; }
         public bool HasBalance { get; set; }
     }
 }
